Improve Wu Jian wrath wrtr barbs/slowness.
For barbs, increase the duration of barbs rather than setting it,
capping it at 12 turns, and set the barbs' power to a maximum of 6,
depending on what it already was set to. These limits are modeled after
those of manticore barbs.

For slowness, use slow_player() to make the player slow for a certain
duration, since that function will handle things properly if the player
is already slowed.

These changes handle the cases where the player already has barbs in
them, and/or is already slowed, when Wu Jian wrath kicks in.