Allow ghouls to heal from all types of enemies
Allow ghouls to heal and become lucid from deaths of non-living, undead
and holy enemies. Ghouls and vampires are the only species in crawl
which have innate traits that get completely denied by the types of
enemies they're facing.

Normally it's fine to have enemies who are immune to certain things
because it forces the player to come up with a new strategy, but in
this case having an enemy say "nope you're worse than a human now" just
feels unfun. Demonspawn powered by death works against non-living and
undead too so it should be fine here for the sake of gameplay.

Give back ghoul's coinflip chance to heal to compensate.