Cap max number of things pulled by Blink Allies Encircling
This used to have a power-based cap on how many allies it could blink at
once. Some of the code even remained to do this, but was rendered
non-functional by fb65b577b5ec8cf810b0ab130fba421e077fe631

Something being accidentally broken in the past is certainly no reason to
return to it, but while this is a great spell that creates great problems,
the possibility of being *completely* surrounded by enemies the instant a
guardian serpent comes into view can sometimes feel a bit excessive. I'm
especially concerned that constriction changes may independently make this
deadlier, so let's try reducing the numbers a little and see how this
feels.

(It now blinks 3-6 units around the player, rather than 'as many as will
fit')