Cap Sticky Flame's duration on players much lower
The cap of 100 turns was leftover from the older incarnation of sticky
flame and not adjusted after 0.31's changes to it. 100 turns is really a
silly high number, and not something one could realistically encounter in
most scenarios, but now that it can do actual relevant damage, it seems
needless to have it be able to stack so high.

I've set the duration at 35 which works out to exactly 5 movements from the
player and is approximately the average duration of being hit by it twice.