Check for flight item removal safety on swap
When removing an item, `_safe_to_remove_or_wear` checked for
the safety of removing the item if it provided flight and the
player was above lava or deep water. However when swapping two
items, `remove` is false and the removal of the old item is
implicit. The function to check the safety of flight removal
`safe_to_remove` is not called in this case.

This commit adds a flight removal safety check in the branch of
`_safe_to_remove_or_wear` that has determined that we're swapping
out an old item, echoing the check that was already happening for
non-swap removal.

I added comments to both `safe_to_remove` checks, because it is
not immediately clear that only flight removal safety is checked by
that funciton. `_safe_to_remove_or_wear` has a comment stating that
it checks for stat zero and does not mention flight, so I wanted
to make sure the code doing something other than what that
comment says was clearly signalled.

Resolves #3502.