Add ARTP_MANA_REGENERATION
Functions the same as the amulet. Amulet of Vitality is now a plain amulet
with 2 levels of ARTP_MANA_REGENERATION instead of being a RegenMP amulet
that was special-cased to count twice.

Amulets of regeneration and magic regeneration used entirely different
methods to track whether the player was attuned to them or not, and this
unifies them, as well as refactoring a whole lot of code related to this.
(This also improves messaging a couple of cases, such as Amulet of Vitality
implying it doesn't work for Djinn at all when you first wear it). I also
reduced some needless scanning of inventory artifact props whenever the
player's regen ticks while they are already at full hp/mp.

There is one functional change here, which is that items which grant both
hp and mp regeneration (currently ONLY the Amulet of Vitality) will no
longer provide either until BOTH of the player's hp and mp are full
simultaneously. This irks me a little, but it seems minor enough that all
of the extra infrastructure required to avoid it (while still allowing
more than a single source of RegenMP to ever function at once) seems
unnecessary.