Properly handle monster halo/umbra from gear.
If a monster's wielding an item changes its halo and/or umbra radius,
invalidate the agrid so that the halo/umbra is drawn properly on monster
placement instead of after monster movement turn afterward (so that e.g.
if Frederick gets an orb of light, it'll work from the beginning).

It's crude, but the functions that accomplish this on the player side
are all completely player-specific.