Prevent allies from displacing you with Electrolunge (lici)
Even if your allies were just trying to be helpful, randomly blinking the
player if you got in their way had the chance of being needlessly
endangering. So now allies won't cast it if they would steal the player's
spot.

(But also add a message if the player somehow gets displaced by the spell
anyway, which is currently narrowly possible (ghosts) and may be more
possible in the future.)