Change Cloak of the Thief's effect
It seems widely agreed these days that Cloak of the Thief's fog effect is
not only double-edged but perhaps actively bad. Sometimes it saves you, but
many other times it reduces your ability to use ?blinking without actually
benefitting you at all. Many experienced players think it may be worse than
a plain cloak and that feels like a sad place for an unrand to be.

So let's try something else entirely!

Instead of any fog effect, Cloak of the Thief now provides noise
suppression while wearing it (equivalent to, and not stacking with, the
mutation). Since this is a very powerful effect, I have also reduced the
enchantment to +0 (and increased the slaying penalty to -3). The hope is
that this is more exciting than the cloak used to be without being
unreasonable. (If I've overshot the numbers, they can always be changed)