refactor: Move some defines to a different header
The define keys related to regular acquirement as well as oka weapon
gifting were in religion-enum.h, when none of them are enums and only
some of them are even related to relgion. Move these to acquire.h,
adding a couple includes in some cc files and removing some includes of
religion-enum.h in a couple other header files. Overall this should
reduce header dependency, since one of the headers with a removed
include is player.h