Don't crash when autodazzler triggers while you're invisible
(And lacking see invis)

The code here somewhat reasonably seems to assume that if there's an agent
with a non-player ktype that they're a monster. But I currently can't seem
to have the dazzler tracer both count friendlies properly *and* not prompt
the player without using the player as its source and using a non-player
ktype.

This feels wrong, and like the real solution lies elsewhere, but for now
let's at least double-check that the agent is a monster instead of
crashing.