Prevent Coglins from getting jewellery-focused troves (Flugkiller)
After the species' restriction of rings in 54e5184 and amulets in ae3e191,
there is no rewards for said species to be found in the troves that solely
place jewellery, so those troves can be restricted for Coglins.

Strictly speaking, nicolae_trove_octotrove places non-jewellery equipment
alongside the amulets and rings. However, shields are also extremely
discouraged for Coglins, which leaves the trove a source of three hats. It
thus has also thus been restricted.