Deathbed conversion clears vengeance (#3576).
For ordinary orcs, as is done for defeated orc apostles (except that
sneak attacks also work on ordinary orcs). Note that this is the only
type of conversion that works on orcs marked for vengeance now.

Add conversion types for vengeance and follower vengeance, and a few
vengeance speech lines.