Don't let players trigger spiders_nest_spider_trap as many times as they want
831329d5 placed a second copy of the same basket of spiders trap in this
vault, which ended up meaning that when you stepped on one of the two
traps, the other trap would become unremovable and you could just
repeatedly trigger it to kill as many (experience and piety-giving)
redbacks as you wanted.

This commit makes it so that only one of the traps actually places; there
are presumably ways to make the vault work with multiple copies but this
seemed like the easiest fix for now.