Tweak weighting in talisman acquirement
Most talismans aren't randarts and thus are useless if they are duplicates,
so increase the weight scaling factor for not-already-seen talismans.

Also, if you have the skill for a talisman of a given tier, that probably
means you already have at least one talisman of that tier and are most
interested in a talisman of tier one higher, not a second talisman of
the same tier (and a talisman of lower tier is even less likely to be
interesting).