Remove a magic breakpoint in book acquirement
A character with no magic skills has a 0% chance of getting a spellbook
instead of a manual from acquirement. Previously the chance of getting a
spellbook increased up to 40% as the proportion of a character's skills
that were magic increased up to 25%, and then suddenly the chance jumped
all the way to 100%. This commit removes the breakpoint, so magic-oriented
characters who previously had a 100% chance of getting a spellbook will
now get a mix of spellbooks and manuals.