Expand the inventory of the One-of-a-Kind Rings shop
We here at the One-of-a-Kind Rings corporate offices would like to
thank you all for your continued support of our franchise. We strive
every day to provide dungeon-goers with the finest of unique rings,
and with the recent advances in artefact-specification technology,
we're pleased to present an all-new addition to our inventories.

One-of-a-Kind Rings franchises will now generate one guaranteed
randart ring with a property that's usually unavailable on randarts.
These properties are:
    Repel Missiles, Clarity, rMut, SH+5, HP+X, and MP+X
    (where X for HP and MP can vary from +10 to +15; for MP, this
    is higher than the usual fixed value of MP+9).

Once again, thank you for your continued patronage.