fix: WebTiles Health and MP spacing (particleface)
The changes in 8f7e8e6e inadvertantly removed the whitespace that comes
after the "HP:" and "MP:" stat labels in the HUD. This was a result of
incorrectly joining the span tags for these labels to the logical lines
of the span tags containing the numeric values portion of the display
for these stats.

This commit adds the new logical lines where necessary. It also removes
some newlines added in 8f7e8e6e that were inserted within the `<span>`
tag elements to allow for use of newlines without making new logical
lines. It's not that hard to work a few overly long lines, and having a
newline inserted after e.g. `<span` and before `>` looks extremely
dubious.