Mark some mutations as evil.
* Black Mark because one of its effects is draining.

* Demonic Magic because a menacing aura that paralyses enemies and is
already described as demonic seems evil.

* Demonic Guardian because we don't know whether the guardian has evil
powers, and filtering all guardians with them will rule out many of the
high-level ones.

* Demonic Touch because it is described as involving unholy energy,
even if it only provides a vulnerability effect.

* Hurl Damnation because damnation is evil.

* Devour On Kill because thriving by killing the living is evil.

* Foul Stench because, while miasma resistance is not evil, poisoning
others with miasma clouds is, and the two effects are tied together.

* Foul Shadow because, while having an umbra is not evil, damaging
others with foul flame is, and the two effects are tied together.

* In Touch With Death because it's a necromancy enhancer, and necromancy
is evil.

* Powered By Death because it's described as involving control over
surrounding life forcees.

* Sanguine Armour and Ignite Blood because blood magic is evil.

* Vampiric because vampirism is evil.