Make abyss exit spawning interrupt travel
At least in local tiles, an abyss exit spawning does not interrupt
autoexplore so you could press o and immediately find yourself 50 steps
away from an exit that just spawned even with a force more for the
message. This commit makes it that having an abyss exit or downstairs
spawn interrupts autoexplore.