Unbug throwing nets
Once upon a time there lived two bugs in a cottage.

The first bug caused a third throwing net that nets a target that's
been netted and then netted again with a second net to get stuck on the
target without increasing the net duration but remain stuck after the
target breaks free, leaving behind a net that catches nobody but yet
cannot be picked up.

The second bug caused netting a target that's standing on top of a net
to merge the nets together, breaking both nets once the target breaks
free.

Now the cottage is on fire and both bugs are dead. Throwing nets that
hit a netted target will not net the target and nets that successfully
net a target will not merge with any nets on the floor.