Improve colours of spells in spellbooks
Previously, in WebTiles, the colours of spells were far from useful.
The most important question when x-v-ing a book and seeing its spells
is "do I have this spell in my library?". There was previously no
way to check this without leaving the x-v menu and manually checking.

Now, all spells will be greyed out in book descriptions if you already
have them in your library (or if they are useless, or if you have
memorised them via Vehumet), lightgrey if eventually useful, and
lightblue if you could memorise it immediately.