Rework and bring back sword of the doom knight
Replace pain brand with the doom brand which causes it to inflict up to
10% of its victim's maximum health on hit. This does the opposite of
the sword of power, dealing more damage the tougher the foe you face
and avoids making you train necromancy for bonus damage when the weapon
has -cast.

Give it ^Drain to discourage swapping it in and out to use spells
because of -cast or only when you face a tough enemy like a unique
panlord.

Also make it evil because doom sounds evil and it was used by a yred
worshipper.