Add punctuation to collisions, remove from misses
It's unclear to me why misses call attack_strength_punctuation when the
damage is always 0, so always use a period instead.

Make punctuation marks in collision messages vary with amount of damage
dealt. This is averaged for collisions between creatures. Collisions
can sometimes deal surprising amounts of damage and this better conveys
that!!