Rework and bring back Crown of Eternal Torment
Replace its mhp malus with *Torm which gives a 5% chance to torment you
and surrounding living creatures when you take non-torment damage. This
serves to give it a more interesting effect and drawback that also
works before extended.

Replace torment immunity with half torment damage so *Torm is an actual
drawback unless you're already undead.

Remove SInv because it doesn't quite fit thematically and we already
have a good number of hat unrands with that.

This commit also adds KILLED_BY_TORMENT as a kill type which is used to
prevent the crown from recursively triggering. It should never actually
kill anything.