Remove AC, EV, SH breakpoints
Previously raising Armour, Dodging, or Shields skill purely for the
defensive benefits (as opposed to decreasing penalties of armour/shields)
did nothing except at discrete points where it suddenly gave you one extra
point of AC (or EV etc). With most skills we try to avoid this sort of
breakpoint so that players gain less benefit from micromanaging skill levels.

Although AC, EV, SH were all already computed to high precision, an internal
value of 8.37 AC was rounded to 8 AC before being used for anything. This
commit changes that - now a value of 8.37 AC means that 8 AC will be used
63% of the time and 9 AC will be used 37% of the time. This type of random
rounding is already used to smooth out the effects of skills in various
other points, including the effect of weapon skill on weapon delay.

The precise internal value of 8.37 AC is currently only visible to the
player when examining an equipment item and seeing how swapping it in will
change their AC. Previously the game might have said "Your AC would decrease
by 1 (8 -> 7)" whereas now it might say "Your AC would decrease by 0.61
(8.37 -> 7.76)". The display of player AC/EV/SH everywhere else is rounded
down to an integer, similar to how weapon delay is displayed rounded down
to a multiple of 0.1. I think there's a decent argument that both weapon
delay and AC/EV/SH (after this commit) should be displayed with one extra
digit of precision "normally", but that can be a separate discussion.

A note on balance: in general I expect this to only have very minor effects.
Previously AC was rounded down to an integer, EV was rounded up,
and SH was rounded to the nearest integer. This means that for a "random"
character not paying attention to the breakpoints, on average this commit
will increase their AC by 0.5 and decrease their EV by 0.5, while leaving
SH unchanged. For a player paying close attention to both AC and EV
breakpoints, this commit will decrease their EV by 1 when they are at both
breakpoints, but help their AC and EV relative to that when they are
training between breakpoints. If there are any concerns about this being a
(minor) player nerf I'd suggest increasing EV by 0.25 or something like that.