Add SH preview and standardise AC/EV/SH previews in descriptions
Previously, the AC change shown in item descriptions included
temporary effects such as Qazlal's AC+3 status or Ignis' fiery armour,
but the EV change did not include temporary changes.

This commit changes the AC preview to match that of EV
in not including temporary effects, and adds an SH preview, which also
does not include temporary effects. I figure that this is the most
consistent strategy, but if it is too confusing, we can always go the
other way. The most important thing is that it is consistent between
AC/EV/SH.

Additionally, with AC/EV/SH now allowed to take fractional values,
all items that affect AC/EV/SH will show the change in their
descriptions, except for non-artefact rings of protection/evasion
and amulets of reflection. E.g. shields of protection will now also
show the AC change, for completeness, and artefact rings of protection/
evasion and amulets of reflection will show their messages (because
strength and dex changes can affect flat bonuses on these items).

I have merged the old method of computing the AC change into the
more recent approach of fake-equipping the item, to simplify some code.