Adjust gravitas pull distance and radius
As far as I can tell with the old formula, the attraction distance for
gells would be always be 1 at radius 1 and 3, 1 at radius 2 below 60
spellpower and 2 otherwise which is pretty unintuitive and has a weird
breakpoint.

Change the distance to 1.25 scaling to 2.5 at max spellpower with
random rounding and remove distance from the center of attraction from
the formula. This does not affect collision damage.

Also increase gell's radius to 3 at 60 spellpower instead of 80.