Use plaintext for more things in species and background YAML files
Change the background YAML files to use plaintext for the
recommended species and skills fields, and the species YAML files
to use plaintext for the recommended jobs, flags, undead, and
recommended_weapons fields (for consistency).

The only enum values remaining are the species' corresponding monsters
and the backgrounds' starting spells, which probably ought to remain
enum values as they don't always correspond to their in-game names.

Hardcode job-gen.py to allow writing simply 'draconian' in the yaml
recommended species field instead of 'base draconian', where the enum
value is SP_BASE_DRACONIAN.