Give lesser beckoning a ministun
This makes using lesser beckoning on melee monsters less bad compared
to just waiting for them to walk next to as you no longer spend 2 MP
just to give them a free turn.

This also makes it consistent with IMB and buffed Gell's ministunning
monsters.

It's fine for spells to be situational and this doesn't change that.