Rework antimagic brand
Simplify antimagic's chance to prevent a spellcast. It is now a flat
2/3 to fizzle a spell rather than 4/(4 + duration) to fail. Having its
effect scale with duration means it is hard to evaluate its usefulness
as a monster with antimagic could either have only 1 turn left and thus
have an 80% chance to succeed casting or enough to never succeed. Since
duration is tied to damage dealt, it also means severely antimagicked
monsters are those who are 1-2 hits away from dying and will die before
antimagic does much and those with high health have so little antimagic
that they will succeed at casting anyway despite the little icon.

Antimagic duration will have a minimum of 10 auts when applied. This
prevents situations where the "magic leaks into the air" message is
printed but no antimagic is actually applied and slightly improves its
performance with low damage weapons.

Antimagic halves your mp rather than reducing it by 2/3. This is still
debilitating enough to casters that they would never want to use it,
but allows non-casters a bit more mp for using god abilities.

Antimagic-branded weapons also gain a huge bonus to accuracy. This is
flavored as antimagic being attracted to latent magic present in all
beings the same way antimatter is attracted to matter.

Since antimagic only works against a small proportion of monsters you
face, and even then likely not a monster with a spell you would care to
disable or one that would survive being hit for more than a couple of
turns, it is only valuable as a swap brand. Players with an antimagic
weapon would either need to swap it frequently with a main weapon to
get value out of it or get so annoyed that they just stick to their
main weapon or rebrand the antimagic weapon. It is also arguable if
disabling a single spellcaster is better than simply killing them
faster when facing multiple enemies is common and antimagic only
disables the enemy you are currently damaging.

Frequent weapon swapping is definitely something we need to move away
from encouraging, so I'm giving antimagic an effect that enhances its
overall effectiveness. This also makes applying antimagic more
consistent and reduces situations where antimagic wears off despite you
making constant attacks against an enemy.