Fix an issue with randart staves in local tiles
When an item that has a randomized appearance (potions, staves) on the
floor is rendered, the item's main tile is actually the "brand". The
randomized tile gets added on later by calling add_main_tile which
calls tileidx_known_base_item on the main tile when. This is a hack,
but it lets you see the brands of identified items on the ground.

However with randart staves, the main tile gets put in the special tile
slot instead. Normally this won't cause any issues, but the local tiles
inventory renders items as if they were on the ground. It calls
add_main_tile on the special slot which then renders the randomized
staff appearance on top of the randart staff tile.

Fixed by not doing this for special tile slots.