fix: reorder some CI steps
Homebrew and system installed python have an unfortunate interaction
from time to time. Most recent github statement on this:

https://github.com/actions/runner-images/issues/9471#issuecomment-1984202712

We already had a workaround in place from c6129fcb3f40, but it is not
working for the current round of the bug. I think reordering should be
the fix. Previously this CI issue must have triggered later in our build
process (currently unclear to me how; maybe the brew calls in the prior
step only recently triggered a dependency upgrade; right now the culprit
is libpng). This commit should ensure it is always triggered before any
brew calls.