fix: further update webtiles tests
It seems that these tests weren't updated when webtiles code was
refactored into a subdirectory. It sort of doesn't matter, because
essentially all of our source code is excluded in tox.ini anyways. (What
is the point of having these tests when everything is excluded? I have
wondered this myself.)