refactor: update stale lock code
This modernizes this code a little bit for python 3, and also checks for
the existence of the process first before prompting the user about
"stale processes". When a server is force-shudown via kill -9,
lockfiles aren't cleared, and the previous version wouldn't actually
check if the pid was real until the 10s prompt had elapsed.