Change the spelling of some words to Australian English.
Change a few places where a word used a spelling which is not spelt
correctly according to
https://www.australian-dictionary.com.au/wordcheck/.

This only changes things in places where the spellings may be displayed
to the user. I include the documentation in this. I haven't changed the
changelog or doc/develop, or any of the contrib stuff.

This changes:

channeling -> channelling

dueling -> duelling

offense -> offence

defense -> defence

skillfull/skillfully -> skilful/skilfully

willful -> wilful