fix: various issues with webtiles idle timers
Since sometime around 2a3aa404fa2d, the play idle timer has been broken.
The issue is essentially race conditions in checking whether a game has
actually started/stopped, where the idle timer was reset during game
startup but before it had fully started. This commit should make both
timers work in a way that is more robust to this issue. The other
problem is that 2a3aa404fa2d seems to have assumed that the connection
life check was only happening during play, which was incorrect. This
lets it work in the lobby, while spectating, and during play again.
The overall implementation is still somewhat painful, because no time is
really tracked in the lobby; possibly it should be.

I also changed the semantics of `max_lobby_idle_time` to match
`max_idle_time` a little better.