Fix coglin weirdness with cursed weapons
It would probably be better to split can_unwield from can_wield, but I
don't want to touch this more than I need to.

The code about trying to equip a two handed weapon as a coglin with
cursed weapons is unnecessary because you are already asked to remove
your off-hand weapon first even if neither is cursed and I don't think
anyone actually wants to use a two handed weapon as a coglin.

Fixes #3562.