Make permissive lof treat actors as obstacles
This change is intended to make ranged attacks and spells much less
annoying to use when you have firewood / allies / slimes blocking the
path by making it choose a slightly different path with the same rules
as permissive los.

Previously when this happened, you would either need to reposition or
drag your targeting behind your target to get a good path that doesn't
hit unwanted targets. This is very annoying as it interrupts tabbing
with a prompt/message and forces you to aim manually.

Currently this may cause some odd behavior with piercing beams.