fix: Bind Soul targeter revealing the true Mara
Yredelemnul's Bind Soul ability checks if the monster was summoned and
in the case of Mara Summon, this allowed the targeter to show which Mara
was not an illusion. Ditto for rashaka's phantom mirror spell.

Fix this by ignoring if the monster is summoned for interface purposes
when the monster has the phantom mirror enchant (currently the targeter
and the displayed potential hp of the bound soul).

Also blacklist player illusions from this effect since this new corner
case, as far as I know, is the only monster with the phantom mirror
effect that is a trivially deducible summon. Is there a better way to do
this?

[Committer's note: slightly edited a comment for clarity]