feat: webtiles-internal https redirect option
This option lets you e.g. redirect http://server:8080 to
https://server:8443. There's various external ways of doing this, but
it's convenient when running webtiles on custom ports to simply be able
to set the redirect in webtiles options.

Note: after spending a bunch of time on this, I've concluded it's
essentially completely impossible within tornado to redirect a http to
https connection on the same port (or vice versa). It's a bit
unfortunate, as this would help with a bunch of scenarios that browsers
auto-manage when using 80/443, but it is what it is. See
https://github.com/tornadoweb/tornado/issues/523 for more info. Also,
tornado just hangs when you send a https connection to a http port,
apparently before the Application code gets access; I have no idea why
this is. But probably official servers will soon no longer use http
ports at all, so maybe not a worry.