Maxwell's Portable Piledriver (new level 4 Translocations)
Intended as a replacement for Manifold Assault as a lower level combat
oriented translocations spell (that scales less well, but needs less combat
skill to function).

It automatically grabs an enemy in melee range, and propels both it and
the caster in that direction until hitting another monster or a wall (or
reaching max range). It does damage via collision only, but that damage
is increased based on how far you've travelled.

Targeting is automatic, and it will pick the enemy who has the longest
unobstructed path to another enemy or wall (to prevent it being far too
commonly unusable with multiple adjacent enemies, where most available
targets are unproductive).

The caster gets -move after using it, proportional to the distance
travelled, and it cannot be recast while you have -move. This is to make it
primarily useful for offensive purposes and short-term repositioning
without being a good way to run away from things (it should always be a net
reduction in long-distance movement speed).

I realize this is similar to Vhi's in several ways, but rather than being
redundant with it, I think they actively complement each other. Piledriver
can target only enemies which Vhi's cannot, moves enemies when Vhi's only
moves the player, and they combo well together - being able to Vhi's into
an enemy to then piledrive it into a wall, or using Vhi's to escape after
doing so (since it bypasses -move)