Move Manifold Assault to level 7, attempt to buff accordingly
Manifold Assault is a fun and popular spell, and having spells that
interact directly with weapon combat is good, but the combination of how
well it scales in lategame and how little investment is needed for many
characters to pick it up by that point is somewhat flattening - it's an
easy and powerful tool for a lot of characters to splash and there's little
reason for those characters ever not to (especially since spellpower
mattered very little for it, in practice)

This commit makes Manifold Assault a lot more of a commitment to get
online, but ideally leaves the spell still feeling powerful and rewarding
when one does so. Compared to its previous incarnation, it now hits twice
as many targets per cast (and scales to hit even more at >100 power), and
can never miss. If this leaves it insuficiently exciting at its new level,
additional buffs will be considered. (I have retained the 'unarmed attacks
hit 1/2 as many things' for now, though I'm not sure whether or not that's
necessary at level 7?).

(Autumn Katana still only hits 4 things at once with its proc, though those
hits will now never miss)