fix: Solve the Mara Hall problem
(DracoOmega named this the "Mara Hall problem" and I love that so I'm
using it here.)

Before this commit, when Mara summoned two illusions of himself, he
would have a 1/2 chance of swapping positions with each illusion
consecutively. This led to a spoilery outcome where you had better odds
of attacking the real Mara if you attacked one of the Maras at the new
positions as demonstrated below:

Call Mara's original position 'a', the position of the first illusion
'b', and the position of the second illusion 'c'.

1st action 2nd action Result position
swap(1/2)  swap(1/2)  c(1/4)
dont(1/2)  swap(1/2)  c(1/4)
swap(1/2)  dont(1/2)  b(1/4)
dont(1/2)  dont(1/2)  a(1/4)

You can see here that the probabilities of the true Mara being at
position a and b are both 1/4 while the sum of the probabilities of Mara
being at position c are 1/2. I don't know of any way position c and b
can be distinguished, but you can usually distinguish position a because
that's where Mara starts. Knowing this, you know that you have equal or
greater odds of attacking the true Mara if you switch from attacking the
Mara at the original position to one of the ones at the new positions.
Like with the Monty Hall problem, you should always switch!

This commit keeps the 1/2 swap chance for the first illusion but uses
1/3 chance for the next potential swap. (and 1/4, 1/5, 1/6, etc for
subsequent swaps but Mara only makes two clones so that never happens in
practice) This evens out the odds as demonstrated by the new table:

1st action 2nd action Result position
swap(1/2)  swap(1/3)  c(1/6)
dont(1/2)  swap(1/3)  c(1/6)
swap(1/2)  dont(2/3)  b(1/3)
dont(1/2)  dont(2/3)  a(1/3)

Now the two c's sum to 1/3 and all three positions have an equal chance
of containing the true Mara!

I don't actually have a proof for the general case of any number of
swaps but simulations suggest that it is true. It seems like something
relevant enough to computer science that there's probably an existing
proof out there. Let me know if you know it!