Add stasis, digging, almost no armour, no body armour mutations
Replace the old fake mutations for formicids' stasis and digging
abilities, and octopodes'/draconians' armour restrictions, with
real mutations. This allows them to be described in detail in
the mutation description.

This leaves species::bans_eq in a rather sorry place, but it
still functions so we'll deal with it later.