Add poison immunity mutation
Replace the old nonliving/undead check for poison immunity,
and the special-cased, weirdly behaving fake mutation for it.
Ghouls, mummies, gargoyles and djinn start with this mutation.
Bloodless vampires remain special-cased.

The poison immunity mutation also implies miasma and sickness
immunity, and this is added to the text of the mutation, rather
than having a separate fake mutation for this that, inconsistently,
appeared for gargoyles and djinn but not for undead species.
[Sidenote: The only source of rPois+++ that doesn't also grant
rMiasma is the staff of Olgreb; perhaps this should be changed---
after all, in this case, miasma only causes slowing anyway.]

Due to the special cases of bloodless vampires and petrification
respectively, I have not removed the undead/nonliving checks from
the poison/miasma immunity functions.

Also, fix the form fake mutation for poison immunity not showing up
if you had a mutation that gave you rPois.