Add amphibiousness as a mutation
Replaces the old convoluted approach. Barachim and octopodes start
with this mutation, and grey draconians now develop it at XL 7, giving
them a message to indicate their new swimming ability.

It might be interesting in future to open this up to some/all species
as a random mutation.

I am leaving in the special-casing that allows merfolk to benefit from
amphibiousness with MUT_MERTAIL because it remains the case that
the text description for the amphibious mutation may be confusing when
merfolk's nimble swimmer 2 is strictly stronger. It would probably
be more consistent to let MUT_AMPHIBIOUS do the work, though.