Mildly brighten up the obsidian bat tile
Various people have complained the tile doesn't contrast enough against the
relatively dark floors of the Orcish Mines. While obsidian bats are kind of
inherently a dark material with a near-black other use of it currently
present, a little bit of lightening this up should help without losing
much of the flavour.