A few more monster-based skill titles (yrdzrfxndfvh), some tidying
With some minor adjustments, this takes some of the titles that weren't
weaker monsters, overly jokey in tone, or weird implications of changing
physiologies on non-Draconians, from #3712.

 * 27 Fighting Humans of Makhleb - the Hell Knight
 * 27 Unarmed Octopodes - the Crusher
 * 27 Armour Trolls - the Iron Troll
 * 27 Shields Spriggans - the Defender
 * 27 Fire Magic Gargoyles - the Molten
 * 27 Invocations Draconians of TSO / Ely / Zin - the Pearl Dragon

These should be relatively straightforwardly understandable each. "Crusher"
is a little silly still as a reference to a removed monster, but it's
probably fine in the same way all the Felid unarmed titles shift away from
the regular unarmed titles also.

(While I'm here, I also sorted skills.cc's list of special titles to follow
the in-game display order of skills and the historical / enum order of when
species were added, rather than the orders of whenever people added these.)