Refactor a couple spell animations
A number of spells that flashd an individual tile as part of their
animation (eg: scorch) were using mostly-duplicated code to do so. This
pulls that out into a separate function so that it's easier to reuse.

(There are a couple other places that could *maybe* make use of this, too,
but these were the most simple and obvious)