Add sticky flame to Hints Mode (PleasingFungus)
While moving around or moving into water is already mentioned in the
earliest user's description, the status's descriptions, and all of the
player and monster spells' descriptions, we clearly haven't emphasized it
enough for many people to have noticed the change. While it'd be quite
intrusive to mention it whenever one's lit on fire (since it'd bleed
across multiple console lines), people in hints mode are already getting a
large amount explanation attached to various matters, and bombardier
beetles are quite early, so, it's probably a good idea to make sure at
least those newcomers get told about this very directly.