Gell's Gavotte (new level 6 Translocations spell)
A different take on gravity manipulation, intended to provide a fun and
novel tool between level 4 and the new climb to level 7 Manifold Assault.
This spell lets the caster temporarily reorient gravity in a cardinal
direction of their choices, moving *everything* that isn't tied down 4
tiles in that direction, potentially falling into walls and each other.

'Everything' includes the caster, so it can be used for repositioning in
some situations, but keep in mind that enemies will usually follow in
lockstep with you unless you cast it from a position where they will bump
into walls instead of being able to follow you.

Enemies only take collision damage if they move at least one space from
this effect (so no bashing things against the sides of a 1-tile hallway)
and there is a several cooldown after casting it to prevent using it
repeatedly to leap across long distances.

Nevertheless, it's a tool that is multipurpose - repositioning yourself,
doing damage, lining up enemies together for other spells. And one that
does so via unique mechanisms that will hopefully be fun to play with.

(The targeter helpfully hilights which monsters will take collision damage
when the spell is aimed in a given direction, rather than asking the
player to count distances themselves)